# Optimize the final distribution bundles