--- a/apps/admin-portal/src/components/index.ts
+++ b/apps/admin-portal/src/components/index.ts
@@ -16,7 +16,6 @@
  * under the License.
  */
 
-export * from "./authentication";
 export * from "./certificates";
 export * from "./claims";
 export * from "./email-templates";