--- a/apps/admin-portal/src/pages/authentication/sign-in.tsx
+++ b/apps/admin-portal/src/pages/authentication/sign-in.tsx
@@ -31,7 +31,7 @@ import { handleSignIn } from "../../store/actions";
  * @param props - Props injected to the component.
  * @return {React.ReactElement}
  */
-export const SignIn: FunctionComponent<RouteComponentProps> = (
+const SignIn: FunctionComponent<RouteComponentProps> = (
     props: RouteComponentProps
 ): ReactElement => {
 
@@ -79,3 +79,10 @@ export const SignIn: FunctionComponent<RouteComponentProps> = (
 
     return null;
 };
+
+/**
+ * A default export was added to support React.lazy.
+ * TODO: Change this to a named export once react starts supporting named exports for code splitting.
+ * @see {@link https://reactjs.org/docs/code-splitting.html#reactlazy}
+ */
+export default SignIn;