--- a/apps/admin-portal/src/pages/authentication/sign-out.tsx
+++ b/apps/admin-portal/src/pages/authentication/sign-out.tsx
@@ -40,3 +40,10 @@ export const SignOut: FunctionComponent<{}> = (): ReactElement => {
 
     return null;
 };
+
+/**
+ * A default export was added to support React.lazy.
+ * TODO: Change this to a named export once react starts supporting named exports for code splitting.
+ * @see {@link https://reactjs.org/docs/code-splitting.html#reactlazy}
+ */
+export default SignOut;