… apps/admin-portal/package.json
… apps/admin-portal/src/pages/authentication/index.ts
… apps/admin-portal/src/pages/certificates/certificates-keystore.tsx
… apps/admin-portal/src/pages/certificates/certificates-truststore.tsx
… apps/admin-portal/src/pages/claims/claim-dialects.tsx
… apps/admin-portal/src/pages/claims/external-dialect-edit.tsx
… apps/admin-portal/src/pages/claims/local-claims-edit.tsx
… apps/admin-portal/src/pages/claims/local-claims.tsx
… apps/admin-portal/src/pages/configurations/server-configurations.tsx
… apps/admin-portal/src/pages/email-templates/email-locale-add.tsx
… apps/admin-portal/src/pages/email-templates/email-template-types.tsx
… apps/admin-portal/src/pages/email-templates/email-templates.tsx
… apps/admin-portal/src/pages/errors/404.tsx
… apps/admin-portal/src/pages/errors/unauthorized.tsx
… apps/admin-portal/src/pages/groups/groups.tsx
… apps/admin-portal/src/pages/index.ts
… apps/admin-portal/src/pages/overview.tsx
… apps/admin-portal/src/pages/privacy.tsx
… apps/admin-portal/src/pages/roles/role-edit.tsx
… apps/admin-portal/src/pages/roles/role.tsx
… apps/admin-portal/src/pages/user-stores/user-stores-edit.tsx
… apps/admin-portal/src/pages/user-stores/user-stores.tsx
… apps/admin-portal/src/pages/user-stores/userstores-templates.tsx
… apps/admin-portal/src/pages/users/user-edit.tsx
… apps/admin-portal/src/pages/users/users.tsx
… apps/admin-portal/src/types/imports.d.ts
… apps/admin-portal/webpack.config.js
… apps/developer-portal/package.json
… apps/developer-portal/scripts/copy-themes.js
… apps/developer-portal/scripts/post-build.js