## Purpose
When the portals are built for the development environments, the `main.js` bundle is oftentimes quite large and it gives bad UX when the user logs in the first time.

## Goals
With this PR, we will try to achieve the following goals.
1. Split the code into multiple chunks and load them on demand.
2. Optimize Webpack configuration.
3. Refactor how we use the assets inside the theme module.

Fixes #1010

## Approach

Following set of optimizations we done in the Webpack configuration files.

1. Use sourcemaps only in development environments.
2. Enable creating runtime chunks apart from the vendor and main chunks.
3. Add hashes to chunks to improve caching.
4. Bundle only supported `moment.js` locales. 
2. Add `gzip` & `brotli` compression support for the prod build.

The following optimizations and refactoring has been done on the mentioned modules.

1. Remove all the references to the `@wso2is/theme` module from inside the portals.
2. Persist a copy of the theme module assets in the portal source to improve tree shaking.
3. Lazily load route components.
4. Stop packing JS version of translations in `@wso2is/i18n` module.
5. Make `@wso2is/react-components` independent from the `@wso2is/theme` module.

## Results

|                               |Main Chunk Size  | Vendor Chunk Size   | Initial Load Time  |
|---|---|---|---|
| User Portal           | 4.7 MB -> 874 KB [ `343 KB gzip` ] |8.9 MB -> 2.8 KB [ `708 KB gzip` ]  | 6743 ms -> 2256 ms  |  
|  Admin Portal        | 5.2 MB -> 1.2 MB [`414 KB gzip`] | 9.2 MB -> 2.8 MB [`708 KB gzip`] | 3976 ms -> 2962 ms  |  
| Developer Portal  | 5.1 MB -> 1.1 MB [`404 KB gzip`]  | 8.8 MB -> 2.5 MB [`644 KB gzip`]  | 4083 ms -> 3073 ms  |  

## TODO
1. Move dependencies to relevant app/module rather than keeping in the root package.json. NOTE: Evaluate Yarn workspaces.
2. Change `lodash` -> `lodash-es` for size optimizations.